# api/docs/swagger.yaml
basePath: /v1
definitions:
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  models.BooksForList:
    properties:
      author_first_name:
        type: string
      author_id:
        type: integer
      author_last_name:
        type: string
      book_type:
        type: string
      category_id:
        type: integer
      id:
        type: integer
      image:
        type: string
      like_count:
        type: integer
      sub_category_id:
        type: integer
      title:
        type: string
      view_count:
        type: integer
    type: object
...
  models.LoginReq:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
  models.UserLoginRes:
    properties:
      access_token:
        type: string
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      login:
        type: string
      password:
        type: string
      role_id:
        type: integer
    type: object
  models.UserReq:
    properties:
      avatar_url:
        example: https://shorturl.at/akpDK
        type: string
      full_name:
        type: string
      login:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      role_id:
        example: 3
        type: integer
    type: object
  models.UserRes:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      like_books:
        items:
          $ref: '#/definitions/models.BooksForList'
        type: array
      login:
        type: string
    type: object
...
  models.UserUpdateReq:
    properties:
      avatar_url:
        type: string
      full_name:
        type: string
      id:
        type: integer
      login:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Book store API
  version: "1.0"
paths:
...
  /client/{id}:
    delete:
      consumes:
      - application/json
      description: Here user can delete info
      operationId: deleteuser user
      parameters:
      - description: DeleteUser
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Here user can get
      operationId: getuser user
      parameters:
      - description: GetUser
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: get user
      tags:
      - User
  /client/login:
    post:
      consumes:
      - application/json
      description: Here user can login
      operationId: loginuser user
      parameters:
      - description: LoginUser
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserLoginRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: login user
      tags:
      - User
  /client/register:
    post:
      consumes:
      - application/json
      description: Here user can register
      operationId: register user
      parameters:
      - description: RegisterUser
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserReq'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserLoginRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: register user
      tags:
      - User
  /client/update:
    put:
      consumes:
      - application/json
      description: Here user can update info
      operationId: updateuser user
      parameters:
      - description: UpdateUser
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.UserRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: update user
      tags:
      - User
...
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"